Fix: scan token TTL 500s when SCAN_TOKEN_TTL is set in .env
env() returns a string, and Carbon 3 type-hints addSeconds(int|float), so
config/scan.php's uncast env('SCAN_TOKEN_TTL', 60) makes every /scan-token
request throw a TypeError. It only works today because the var is unset and
the int default is used — set it in .env, as anyone tuning the TTL would,
and the whole QR feature 500s.

Cast in the config and defensively in the service (which protects the value
whatever its source), plus a regression test with a string TTL.

Found during the on-device app-two scanner pass.

+++ app/Services/ScanTokenService.php
@@ -10,7 +10,7 @@ class ScanTokenService
     public function issue(ParentAccount $parent, ?Carbon $at = null): array
     {
         $at ??= now();
-        $ttl = config('scan.token_ttl', 60);
+        $ttl = (int) config('scan.token_ttl', 60);
         $secret = config('scan.qr_secret');
 
         $payload = [
@@ -31,7 +31,7 @@ class ScanTokenService
     public function verify(string $token, ?Carbon $at = null): ?int
     {
         $at ??= now();
-        $ttl = config('scan.token_ttl', 60);
+        $ttl = (int) config('scan.token_ttl', 60);
         $secret = config('scan.qr_secret');
 
         $parts = explode('.', $token, 2);
