+++ config/scan.php
@@ -2,5 +2,5 @@
 
 return [
     'qr_secret' => env('SCAN_QR_SECRET', env('APP_KEY')),
-    'token_ttl' => env('SCAN_TOKEN_TTL', 60),
+    'token_ttl' => (int) env('SCAN_TOKEN_TTL', 60),
 ];
